tbv2: use std::decay_t in make_field
CodeGen v2 permits template parameters to be qualified. This means that if we
call `make_field` with a template parameter it will be qualified. However, we
don't qualify the types when generating meta functions such as `NameProvider`
and `TypeHandler`. This means these qualified types don't match up with the
expected type.

Use `std::decay_t` when forwarding the type to `NameProvider` and `TypeHandler`
so they're always the base type that they were generated with.

Test Plan:
- CI
- Added a test for `std::unique_ptr<const Foo>` to exercise this. Failed
  before, passes after.