circleci: run oid in legacy tests
Currently the CodeGen v1 CI tests are filtered to only run OIL tests. This is
incorrect, as OIL always runs with CodeGen v2. Change it so only OID tests are
run as legacy.

Also restructured the CI file a bit to group the build steps together and the
test steps together. I can revert this if it's inconvenient but I think it
makes it clearer.

Test plan:
- Checked the CI and it's running OID tests now