tbv2: re-add indirection to list
`std::list` was intended to have an extra layer to represent the overhead of
each list element similar to map implementations. This wasn't currently being
output which meant the size of each list element was being undercounted.

Add the extra layer of indirection to list elements and include this in the
tests.

Test Plan:
- CI
- Updated test cases