tbv2: add is_primitive to output
C++ has a concept of Primitive which holds in the type graph. However we don't
currently expose this information to the end user. Expose this from the OIL
iterator to allow future features like primitive rollups.

This affects containers like maps and `std::list` which have a fake `[]`
element with no type. They use this to group together the key/value in a map
and to account for any per element storage overhead. Currently the decision is
to make the fake `[]` element a primitive if all of its children are
primitives. This allows for more effective primitive rollups if that is
implemented. This implementation detail may be changed in future.

Test Plan:
- CI
- Updated simple tests.