circleci: drop
CircleCI is no longer an option for a facebookexperimental repo and our builds
are now failing because they rely on larger runners that are only available with
paid accounts. Drop CircleCI.

We lose coverage by doing this, but the testing is covered appropriately by
GitHub Actions. Will try to follow up with PRs that re-enable coverage.

Test plan:
- CI